requirements.txt
Django>=5.0,<6.0
gunicorn
psycopg2-binary
python-dotenv
